Makefile: Extensive re-organization; add releases: default target
The only functionality change is a new `releases` target at the top of the
file so that the default target now just creates all the release files.

The rest is simply moving the targets around so that each platform gets its
own section to make it easier to understand how each individual platform is
being built (and easier for those who want to hack on just one platform's
build).

Each platform section starts with the "run" target for that platform
(whether on a simulator or via download to the real computer), then the
top-level release target, then any additional targets required by that
platform build.