Makefile clean: re-order files on to a line for each platform
Again, for readability when working on build related to a specific
platform.

An alternative would be to have a separate `clean-apple2` etc. target
for each platform, and have the top level `clean` depend on those, but
that seems more work than is worthwhile at the moment.

Note that the `clean` target isn't entirely correct; there are still a
couple of files left behind and a couple removed that may not actually
be generated. The latter are left on the last line until we figure out
if they are generated even though not explicitly mentioned in the
Makefile.