Makefile clean: Add missing files
As discovered with `git status --ignored`, the `apple2` and `rc6502`
targets were generating some files that were not cleaned.

We add a shell comment to the end of the command line so that we
need not worry about special-casing the last line that actually
does anything not to have an escape (backslash) at the end.