Makefile apple2: Use linapple if available for Linux compatibility
We try to run `linapple --help` and, if it works, we use that; otherwise
we use what was previously the sole method: `open ~/Applications/Virtual ][`.

Obviously Linux users will have to have `linapple` in their path. One could
get a lot more sophisticated about checking which system one is on and
giving appropriate error messages, but that seems more effort than is
really worthwhile.